pyproject.toml
[project]
name = "itdepends"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "numpy==2.2.6",
    "openai[beta]>=1.86.0",
    "pydantic>=2.11.6",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "soundcard>=0.4.4",
    "sounddevice>=0.5.2",
    "soundfile>=0.13.1",
    "websockets>=13.1",
]
